Увеличил шанс показа студента того же года поступления и переделал класс RandomCommandHandler для расширяемости
